fix(admin): payment 订单金额 Long（分）全链对齐——wire/北向 Response/Query 同型零换算（issue #58，#55 T2）
payment 侧金额本就是 Long（分）：PaymentOrderResponse.amount 为 Long、列表 Query 的
amountMin/Max 为 Long。admin 侧 wire 镜像 DTO、北向 Response、北向 Query 却全为
BigDecimal——类型撒谎（Jackson 宽松反序列化掩盖，值上常能跑通），且北向 Query javadoc
错称「金额暂以元（BigDecimal）承」（事实契约一直是分，前端 yuanToCents 先行）。本票
admin 单方面对齐（ADR-0011 §1），payment 零改动：

- 订单列表 + 详情四个 DTO（PaymentOrderWireResponse / PaymentOrderDetailWireResponse /
  PaymentOrderSummaryResponse / PaymentOrderDetailResponse）amount BigDecimal → Long
- 北向 PaymentOrderQuery 与 wire PaymentOrderListWireRequest 的 amountMin/Max 同步
  Long；javadoc 撤销「暂以元承」立场、更正为分（ADR-0011），订单侧已无「元」措辞
- 北向 JSON 金额随之变 string 形态（框架全局 Long→ToStringSerializer，ADR-0011 接受），
  前端联动见 aieducenter-admin-web#54 变更点 1
- 列表信封契约测试 fixture 金额改 string 形态（"amount": "9900"——payment 真实发出
  的形状）+ 新增 Long 透传断言（断在真实 seam 返回的北向 Response）
- BFF 集成测试订单侧 fixture 同步 Long 分（99.00→9900L 等）；退款/stats 侧 BigDecimal
  不动（#59 / #60 后续票）

退款金额对齐（#59）、stats 金额 + refundBacklog 嵌套（#60）另票跟进。

+++ src/main/java/com/aieducenter/admin/payment/application/dto/query/PaymentOrderQuery.java
@@ -1,6 +1,5 @@
 package com.aieducenter.admin.payment.application.dto.query;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,7 +12,8 @@ import java.util.List;
  *
  * <p>枚举筛选项（{@code statuses} / {@code payMode} / {@code accessType} / {@code paymentChannel}）以 payment
  * {@code BaseEnum} 的 <strong>Integer code</strong> 传递——payment 按 code 绑定枚举（cartisan-web BaseEnum↔code），
- * 非 enum name。金额筛选 {@code amountMin/Max} 暂以元（BigDecimal）承，与 payment 的分（Long）对齐见 #55。</p>
+ * 非 enum name。金额筛选 {@code amountMin/Max} 为 <strong>Long（分）</strong>，与 payment 同型透传、零换算
+ * （ADR-0011——前端已按分提交，admin 不做分↔元换算）。</p>
  *
  * @since 0.1.0
  */
@@ -40,11 +40,11 @@ public record PaymentOrderQuery(
         /** 支付通道（payment BaseEnum code） */
         Integer paymentChannel,
 
-        /** 金额下限（含） */
-        BigDecimal amountMin,
+        /** 金额下限（分，含） */
+        Long amountMin,
 
-        /** 金额上限（含） */
-        BigDecimal amountMax,
+        /** 金额上限（分，含） */
+        Long amountMax,
 
         /** 创建时间起（含） */
         LocalDateTime createdAtFrom,
