+++ src/main/java/com/aieducenter/admin/payment/application/dto/response/PaymentOrderDetailResponse.java
@@ -1,6 +1,5 @@
 package com.aieducenter.admin.payment.application.dto.response;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -13,6 +12,9 @@ import java.time.LocalDateTime;
  * <p>枚举出口规则（ADR-0009）：{@code status}/{@code statusName}、{@code payMode}/{@code payModeName}、
  * {@code accessType}/{@code accessTypeName}、{@code paymentChannel}/{@code paymentChannelName}，枚举 code 为 Integer。
  * 前端直读 {@code *Name}，不在端侧做枚举→中文映射。中文名由 payment 出口提供、admin 透传。</p>
+ *
+ * <p>金额出口规则（ADR-0011）：{@code amount} 为 <strong>Long（分）</strong>，与 payment 同型透传、零换算；
+ * 北向 JSON 中呈 string 形态（框架全局 Long→{@code ToStringSerializer}），前端算术入口统一转 {@code Number}。</p>
  *
  * @since 0.1.0
  */
@@ -28,7 +30,8 @@ public record PaymentOrderDetailResponse(
 
         String statusName,
 
-        BigDecimal amount,
+        /** 支付金额（分） */
+        Long amount,
 
         Integer payMode,
